requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# CAD Processing
cadquery-ocp==7.7.2
numpy==1.26.4

# CORS
starlette==0.40.0
